+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.31.4
+* Notification updated
+* More logs added
+
 ## 0.31.3
 * Search by sender implemented in groups
 * Minor improvements